Adopt SszUInt64 for extension
