+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/teku/reference/phase0/ssz_generic/containers/SszUInt16.java
@@ -15,7 +15,7 @@ package tech.pegasys.teku.reference.phase0.ssz_generic.containers;
 
 import tech.pegasys.teku.infrastructure.ssz.impl.AbstractSszPrimitive;
 
-public class SszUInt16 extends AbstractSszPrimitive<Integer, SszUInt16> {
+public class SszUInt16 extends AbstractSszPrimitive<Integer> {
 
   public static SszUInt16 of(Integer val) {
     return new SszUInt16(val);
