+++ infrastructure/ssz/src/main/java/tech/pegasys/teku/infrastructure/ssz/collections/SszMutablePrimitiveVector.java
@@ -17,7 +17,7 @@ import tech.pegasys.teku.infrastructure.ssz.SszMutableVector;
 import tech.pegasys.teku.infrastructure.ssz.SszPrimitive;
 
 public interface SszMutablePrimitiveVector<
-        ElementT, SszElementT extends SszPrimitive<ElementT, SszElementT>>
+        ElementT, SszElementT extends SszPrimitive<ElementT>>
     extends SszMutablePrimitiveCollection<ElementT, SszElementT>,
         SszMutableVector<SszElementT>,
         SszPrimitiveVector<ElementT, SszElementT> {
