+++ infrastructure/ssz/src/main/java/tech/pegasys/teku/infrastructure/ssz/collections/impl/AbstractSszMutablePrimitiveCollection.java
@@ -34,7 +34,7 @@ import tech.pegasys.teku.infrastructure.ssz.tree.TreeNode;
 import tech.pegasys.teku.infrastructure.ssz.tree.TreeUpdates;
 
 public abstract class AbstractSszMutablePrimitiveCollection<
-        ElementT, SszElementT extends SszPrimitive<ElementT, SszElementT>>
+        ElementT, SszElementT extends SszPrimitive<ElementT>>
     extends AbstractSszMutableCollection<SszElementT, SszElementT>
     implements SszMutablePrimitiveCollection<ElementT, SszElementT> {
 
@@ -117,7 +117,7 @@ public abstract class AbstractSszMutablePrimitiveCollection<
   }
 
   private static class NodeUpdate<
-      ElementT, SszElementT extends SszPrimitive<ElementT, SszElementT>> {
+      ElementT, SszElementT extends SszPrimitive<ElementT>> {
     private final List<PackedNodeUpdate<ElementT, SszElementT>> updates;
     private final long nodeGIndex;
