+++ infrastructure/ssz/src/main/java/tech/pegasys/teku/infrastructure/ssz/primitive/SszNone.java
@@ -16,7 +16,7 @@ package tech.pegasys.teku.infrastructure.ssz.primitive;
 import tech.pegasys.teku.infrastructure.ssz.impl.AbstractSszPrimitive;
 import tech.pegasys.teku.infrastructure.ssz.schema.SszPrimitiveSchemas;
 
-public class SszNone extends AbstractSszPrimitive<Void, SszNone> {
+public class SszNone extends AbstractSszPrimitive<Void> {
 
   public static final SszNone INSTANCE = new SszNone();
