+++ infrastructure/ssz/src/main/java/tech/pegasys/teku/infrastructure/ssz/schema/SszPrimitiveSchemas.java
@@ -19,11 +19,7 @@ import static tech.pegasys.teku.infrastructure.ssz.schema.json.SszPrimitiveTypeD
 import static tech.pegasys.teku.infrastructure.ssz.schema.json.SszPrimitiveTypeDefinitions.SSZ_BYTES4_TYPE_DEFINITION;
 import static tech.pegasys.teku.infrastructure.ssz.schema.json.SszPrimitiveTypeDefinitions.SSZ_NONE_TYPE_DEFINITION;
 import static tech.pegasys.teku.infrastructure.ssz.schema.json.SszPrimitiveTypeDefinitions.SSZ_UINT256_TYPE_DEFINITION;
-import static tech.pegasys.teku.infrastructure.ssz.schema.json.SszPrimitiveTypeDefinitions.SSZ_UINT64_TYPE_DEFINITION;
 
-import java.nio.ByteOrder;
-import java.util.Arrays;
-import java.util.List;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import org.apache.tuweni.bytes.MutableBytes;
@@ -31,7 +27,6 @@ import org.apache.tuweni.units.bigints.UInt256;
 import tech.pegasys.teku.infrastructure.bytes.Bytes4;
 import tech.pegasys.teku.infrastructure.json.types.DeserializableTypeDefinition;
 import tech.pegasys.teku.infrastructure.ssz.SszData;
-import tech.pegasys.teku.infrastructure.ssz.SszPrimitive;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszBit;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszByte;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszBytes32;
@@ -40,6 +35,7 @@ import tech.pegasys.teku.infrastructure.ssz.primitive.SszNone;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt256;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt64;
 import tech.pegasys.teku.infrastructure.ssz.schema.impl.AbstractSszPrimitiveSchema;
+import tech.pegasys.teku.infrastructure.ssz.schema.impl.AbstractSszUInt64Schema;
 import tech.pegasys.teku.infrastructure.ssz.schema.json.SszPrimitiveTypeDefinitions;
 import tech.pegasys.teku.infrastructure.ssz.sos.SszDeserializeException;
 import tech.pegasys.teku.infrastructure.ssz.tree.LeafDataNode;
@@ -164,89 +160,10 @@ public final class SszPrimitiveSchemas {
       };
 
 
-  public static abstract class SszUInt64Schema<T extends SszUInt64>
-      extends AbstractSszPrimitiveSchema<UInt64, T> {
-
-    protected SszUInt64Schema() {
-      super(64);
-    }
-
-    @Override
-    public UInt64 createFromLeafBackingNode(LeafDataNode node, int internalIndex) {
-      Bytes leafNodeBytes = node.getData();
-      try {
-        Bytes elementBytes = leafNodeBytes.slice(internalIndex * 8, 8);
-        return UInt64.fromLongBits(elementBytes.toLong(ByteOrder.LITTLE_ENDIAN));
-      } catch (Exception e) {
-        // additional info to track down the bug https://github.com/PegaSysEng/teku/issues/2579
-        String info =
-            "Refer to https://github.com/PegaSysEng/teku/issues/2579 if see this exception. ";
-        info += "internalIndex = " + internalIndex;
-        info += ", leafNodeBytes: " + leafNodeBytes.getClass().getSimpleName();
-        try {
-          info += ", leafNodeBytes = " + leafNodeBytes.copy();
-        } catch (Exception ex) {
-          info += "(" + ex + ")";
-        }
-        try {
-          info += ", leafNodeBytes[] = " + Arrays.toString(leafNodeBytes.toArray());
-        } catch (Exception ex) {
-          info += "(" + ex + ")";
-        }
-        throw new RuntimeException(info, e);
-      }
-    }
-
-    @Override
-    public TreeNode updateBackingNode(TreeNode srcNode, int index, SszData newValue) {
-      Bytes uintBytes =
-          Bytes.ofUnsignedLong(((SszUInt64) newValue).longValue(), ByteOrder.LITTLE_ENDIAN);
-      Bytes curVal = ((LeafNode) srcNode).getData();
-      Bytes newBytes = updateExtending(curVal, index * 8, uintBytes);
-      return LeafNode.create(newBytes);
-    }
-
-    @Override
-    public TreeNode updatePackedNode(
-        TreeNode srcNode, List<PackedNodeUpdate<UInt64, T>> updates) {
-      if (updates.size() == 4) {
-        byte[] data = new byte[32];
-        for (int i = 0; i < 4; i++) {
-          long longValue = updates.get(i).getNewValue().longValue();
-          int off = i * 8;
-          data[off + 0] = (byte) longValue;
-          data[off + 1] = (byte) (longValue >> 8);
-          data[off + 2] = (byte) (longValue >> 16);
-          data[off + 3] = (byte) (longValue >> 24);
-          data[off + 4] = (byte) (longValue >> 32);
-          data[off + 5] = (byte) (longValue >> 40);
-          data[off + 6] = (byte) (longValue >> 48);
-          data[off + 7] = (byte) (longValue >> 56);
-        }
-        return LeafNode.create(Bytes.wrap(data));
-      } else {
-        return super.updatePackedNode(srcNode, updates);
-      }
-    }
-
-    @Override
-    public TreeNode getDefaultTree() {
-      return LeafNode.ZERO_LEAVES[8];
-    }
-
-    @Override
-    public DeserializableTypeDefinition<?> getJsonTypeDefinition() {
-      return SSZ_UINT64_TYPE_DEFINITION;
-    }
-
-    @Override
-    public String toString() {
-      return "UInt64";
-    }
-  };
+  ;
 
 
-  public static final AbstractSszPrimitiveSchema<UInt64, SszUInt64> UINT64_SCHEMA = new SszUInt64Schema<>() {
+  public static final AbstractSszPrimitiveSchema<UInt64, SszUInt64> UINT64_SCHEMA = new AbstractSszUInt64Schema<>() {
     @Override
     public SszUInt64 boxed(UInt64 rawValue) {
       return SszUInt64.of(rawValue);
@@ -358,22 +275,6 @@ public final class SszPrimitiveSchemas {
           return "Bytes32";
         }
       };
-
-  private static Bytes updateExtending(Bytes origBytes, int origOff, Bytes newBytes) {
-    if (origOff == origBytes.size()) {
-      return Bytes.wrap(origBytes, newBytes);
-    } else {
-      final MutableBytes dest;
-      if (origOff + newBytes.size() > origBytes.size()) {
-        dest = MutableBytes.create(origOff + newBytes.size());
-        origBytes.copyTo(dest, 0);
-      } else {
-        dest = origBytes.mutableCopy();
-      }
-      newBytes.copyTo(dest, origOff);
-      return dest;
-    }
-  }
 
   abstract static class SszByteSchema extends AbstractSszPrimitiveSchema<Byte, SszByte> {
