+++ infrastructure/ssz/src/test/java/tech/pegasys/teku/infrastructure/ssz/collections/SszMutablePrimitiveCollectionTestBase.java
@@ -26,7 +26,7 @@ public interface SszMutablePrimitiveCollectionTestBase
 
   @MethodSource("sszMutableCompositeArguments")
   @ParameterizedTest
-  default <ElT, SszT extends SszPrimitive<ElT, SszT>> void setElement_throwsIndexOutOfBounds(
+  default <ElT, SszT extends SszPrimitive<ElT>> void setElement_throwsIndexOutOfBounds(
       SszMutablePrimitiveCollection<ElT, SszT> collection) {
     assertThatThrownBy(
             () ->
@@ -38,7 +38,7 @@ public interface SszMutablePrimitiveCollectionTestBase
 
   @MethodSource("sszMutableCompositeArguments")
   @ParameterizedTest
-  default <ElT, SszT extends SszPrimitive<ElT, SszT>> void setElement_extendsExtendableCollection(
+  default <ElT, SszT extends SszPrimitive<ElT>> void setElement_extendsExtendableCollection(
       SszMutablePrimitiveCollection<ElT, SszT> collection) {
     if (collection.size() < collection.getSchema().getMaxLength()) {
       // collection is extendable (List effectively)
